fix fs

+++ crates/gpui/src/platform/windows/window.rs
@@ -905,7 +905,7 @@ impl RawWindow {
     }
 
     pub fn scale_factor(&self) -> f32 {
-        (self.get_dpi() as f32) / (96 as f32)
+        (self.get_dpi() as f32) / 96.0
     }
 
     fn get_dpi(&self) -> u32 {
